+++ src/main/kotlin/org/wit/vidya/console/models/VidyaStore.kt
@@ -2,7 +2,7 @@ package org.wit.vidya.console.models
 
 interface VidyaStore {
     fun findAll(): List<VidyaModel>
-    fun findOne(id: Long): VidyaModel?
+    fun findOne(id: Int): VidyaModel?
     fun create(vidya: VidyaModel)
     fun edit(vidya: VidyaModel)
     fun delete(vidya: VidyaModel)
